pyproject.toml
[project]
name = "vibe-engineer"
version = "0.1.0"
description = "Documentation-driven development workflow tooling"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "claude-agent-sdk>=0.1.44",
    "click>=8.0",
    "httpx>=0.27.0",
    "jinja2",
    "pydantic",
    "pyyaml",
    "scikit-learn",
    "starlette>=0.36.0",
    "uvicorn[standard]>=0.27.0",
]

[project.scripts]
ve = "ve:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
only-include = ["src"]
sources = ["src"]

[tool.hatch.build.targets.sdist]
include = ["src/**", "README.md"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"
markers = [
    "network: marks tests as requiring network access (deselect with '-m \"not network\"')",
]

[tool.coverage.run]
source = ["src"]

[tool.coverage.report]
show_missing = true

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
]
